bots/sled-commander/requirements.txt
python-telegram-bot>=21.0
sqlite-utils>=3.35
python-dotenv>=1.0.0
httpx>=0.27.0
redis>=5.0.0
